rebase from main and resolved conflict and Updated BookDetailsPage UI
